Release 2(?)
First playable release.
Still needs some tweaking with the ores.
And need to figure out which mods are causing Cave Spelunking to break.
And the caves need to be made more rare.
Harder branch mining still needs to be configured.
And maybe one of those surface ores mod?
Either Factori- ores or that Create addon.